@@ -1,7 +1,7 @@
 [package]
 name = "tc2-agent"
-version = "0.0.0-snapshot"  # Updated by travis on release.
-edition = "2021"
+version = "0.0.0-snapshot"  # Updated by CI on release.
+edition = "2024"
 description = "Program to interface between the Raspberry Pi and the RP2040 in the DOC AI Cam"
 license = "GPL v3.0"
 authors = ["Cacophony Developers <coredev@cacophony.org.nz>"]
@@ -29,12 +29,8 @@ sun-times = "0.2.0"
 louvre = "0.2.1"
 rustbus = "0.19.3"
 sha256 = "1.5.0"
-sysinfo = "0.34.2"
+sysinfo = "0.35.2"
 
-[target.aarch64-unknown-linux-musl]
-# applies to target-specific builds
-# rPi3
-rustflags = "-C target-cpu=cortex-a53"
 
 [package.metadata.deb]
 name = "tc2-agent"
