@@ -1,10 +1,6 @@
-use crate::cptv_header::{decode_cptv_header_streaming, CptvHeader};
+use crate::cptv_header::{CptvHeader, decode_cptv_header_streaming};
 use chrono::{DateTime, Utc};
-use flate2::read::GzEncoder;
-use flate2::read::MultiGzDecoder;
-use flate2::Compression;
 use log::{error, info};
-use std::io::prelude::*;
 use std::{fs, thread};
 use thread_priority::{ThreadBuilderExt, ThreadPriority};
 
@@ -18,20 +14,18 @@ pub fn save_cptv_file_to_disk(mut cptv_bytes: Vec<u8>, output_dir: &str) {
         move |_| match decode_cptv_header_streaming(&cptv_bytes) {
             Ok(header) => match header {
                 CptvHeader::V2(header) => {
-                    info!("Saving CPTV file with header {:?}", header);
+                    info!("Saving CPTV file with header {header:?}");
                     let recording_date_time =
                         DateTime::from_timestamp_millis(header.timestamp as i64 / 1000)
                             .unwrap_or(chrono::Local::now().with_timezone(&Utc))
                             .with_timezone(&chrono::Local);
                     if !fs::exists(&output_dir).unwrap_or(false) {
-                        fs::create_dir(&output_dir)
-                            .expect(&format!("Failed to create output directory {}", output_dir));
+                        fs::create_dir(&output_dir).unwrap_or_else(|_| {
+                            panic!("Failed to create output directory {output_dir}")
+                        });
                     }
-                    let path = format!(
-                        "{}/{}.cptv",
-                        output_dir,
-                        recording_date_time.format("%Y-%m-%d--%H-%M-%S")
-                    );
+                    let filename = recording_date_time.format("%Y-%m-%d--%H-%M-%S");
+                    let path = format!("{output_dir}/{filename}.cptv",);
                     //very slow and cpu intensive offered 18% saving on a 10 minute recording
                     // let decoder = MultiGzDecoder::new(&cptv_bytes[..]);
                     // let mut encoder = GzEncoder::new(decoder, Compression::default());
@@ -50,12 +44,11 @@ pub fn save_cptv_file_to_disk(mut cptv_bytes: Vec<u8>, output_dir: &str) {
                     if !is_existing_file {
                         match fs::write(&path, &cptv_bytes) {
                             Ok(()) => {
-                                info!("Saved CPTV file {}", path);
+                                info!("Saved CPTV file {path}");
                             }
                             Err(e) => {
                                 error!(
-                                    "Failed writing CPTV file to storage at {}, reason: {}",
-                                    path, e
+                                    "Failed writing CPTV file to storage at {path}, reason: {e}"
                                 );
                             }
                         }
@@ -78,13 +71,13 @@ pub fn save_cptv_file_to_disk(mut cptv_bytes: Vec<u8>, output_dir: &str) {
                         //     }
                         // }
                     } else {
-                        error!("File {} already exists, discarding duplicate", path);
+                        error!("File {path} already exists, discarding duplicate");
                     }
                 }
                 _ => error!("Unsupported CPTV file format, discarding file"),
             },
             Err(e) => {
-                error!("Invalid CPTV file: ({:?}), discarding", e);
+                error!("Invalid CPTV file: ({e:?}), discarding");
             }
         },
     );
