@@ -2,13 +2,13 @@ use crate::camera_transfer_state::CameraHandshakeInfo;
 use crate::cptv_frame_dispatch;
 use crate::program_rp2040::program_rp2040;
 use crate::recording_state::{RecordingMode, RecordingState};
-use crate::socket_stream::{get_socket_address, SocketStream};
-use crate::telemetry::{read_telemetry, Telemetry};
+use crate::socket_stream::{SocketStream, get_socket_address};
+use crate::telemetry::{Telemetry, read_telemetry};
 use log::{debug, error, info, warn};
 use rppal::gpio::OutputPin;
+use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::mpsc::{Receiver, RecvTimeoutError};
-use std::sync::Arc;
 use std::thread::sleep;
 use std::time::{Duration, Instant};
 use std::{process, thread};
@@ -24,8 +24,10 @@ fn restart_rp2040_if_requested(
     run_pin: &mut OutputPin,
     restart_rp2040_ack: &mut Arc<AtomicBool>,
 ) {
+    // FIXME: Check if the rp2040 recording flag is set before restarting.
+
     // Check if we need to reset rp2040 because of a config change
-    if let Ok(_) = restart_rp2040_channel_rx.try_recv() {
+    if restart_rp2040_channel_rx.try_recv().is_ok() {
         restart_rp2040_ack.store(true, Ordering::Relaxed);
         info!("Restarting rp2040");
         if !run_pin.is_set_high() {
@@ -47,76 +49,73 @@ pub fn spawn_frame_socket_server_thread(
     recording_state: &RecordingState,
 ) {
     let recording_state = recording_state.clone();
-    let _ = thread::Builder::new()
-        .name("frame-socket".to_string())
-        .spawn_with_priority(ThreadPriority::Max, move |result| {
+    let _ = thread::Builder::new().name("frame-socket".to_string()).spawn_with_priority(
+        ThreadPriority::Max,
+        move |result| {
             let address = get_socket_address(serve_frames_via_wifi);
             let management_address = "/var/spool/managementd".to_string();
             // Spawn a thread which can output the frames, converted to rgb grayscale
             // This is printed out from within the spawned thread.
             if result.is_err() {
                 error!(
-                    "Thread must have permissions to run with realtime priority, run as root user"
+                    "Thread must have permissions to run with realtime priority, \
+                    run as root user"
                 );
                 process::exit(1);
             }
 
             let mut reconnects = 0;
             let mut prev_frame_num = None;
-            let mut sockets: [(String, bool, Option<SocketStream>); 2] = [
-                (address, serve_frames_via_wifi, None),
-                (management_address, false, None),
-            ];
+            let mut sockets: [(String, bool, Option<SocketStream>); 2] =
+                [(address, serve_frames_via_wifi, None), (management_address, false, None)];
+
+            restart_rp2040_if_requested(
+                &restart_rp2040_channel_rx,
+                &mut run_pin,
+                &mut restart_rp2040_ack,
+            );
+            let mut recv_timeout_ms = 10;
+            info!("Connecting to frame sockets");
+            let mut ms_elapsed = 0;
             loop {
                 restart_rp2040_if_requested(
                     &restart_rp2040_channel_rx,
                     &mut run_pin,
                     &mut restart_rp2040_ack,
                 );
-                let mut recv_timeout_ms = 10;
-                info!("Connecting to frame sockets");
-                let mut ms_elapsed = 0;
-                loop {
-                    restart_rp2040_if_requested(
-                        &restart_rp2040_channel_rx,
-                        &mut run_pin,
-                        &mut restart_rp2040_ack,
-                    );
-                    if recording_state.recording_mode() == RecordingMode::Thermal {
-                        for (address, use_wifi, stream) in
-                            sockets.iter_mut().filter(|(_, _, stream)| stream.is_none())
-                        {
-                            let stream_connection: Option<SocketStream> =
-                                SocketStream::from_address(&address, *use_wifi).ok();
-                            if stream_connection.is_some() {
-                                println!("Connected to {}", &address);
-                            }
-                            *stream = stream_connection;
+                if recording_state.recording_mode() == RecordingMode::Thermal {
+                    for (address, use_wifi, stream) in
+                        sockets.iter_mut().filter(|(_, _, stream)| stream.is_none())
+                    {
+                        let stream_connection: Option<SocketStream> =
+                            SocketStream::from_address(address, *use_wifi).ok();
+                        if stream_connection.is_some() {
+                            println!("Connected to {address}");
                         }
-
-                        let connections = sockets
-                            .iter()
-                            .filter(|(_, _, stream)| stream.is_some())
-                            .count();
-                        if connections == 0 {
-                            sleep(Duration::from_millis(1000));
-                            continue;
-                        }
+                        *stream = stream_connection;
                     }
 
-                    handle_payload_from_frame_acquire_thread(
-                        camera_handshake_channel_rx
-                            .recv_timeout(Duration::from_millis(recv_timeout_ms)),
-                        &mut sockets,
-                        &mut ms_elapsed,
-                        &mut reconnects,
-                        &mut prev_frame_num,
-                        &recording_state,
-                        &mut recv_timeout_ms,
-                    );
+                    let connections =
+                        sockets.iter().filter(|(_, _, stream)| stream.is_some()).count();
+                    if connections == 0 {
+                        sleep(Duration::from_millis(1000));
+                        continue;
+                    }
                 }
+
+                handle_payload_from_frame_acquire_thread(
+                    camera_handshake_channel_rx
+                        .recv_timeout(Duration::from_millis(recv_timeout_ms)),
+                    &mut sockets,
+                    &mut ms_elapsed,
+                    &mut reconnects,
+                    &mut prev_frame_num,
+                    &recording_state,
+                    &mut recv_timeout_ms,
+                );
             }
-        });
+        },
+    );
 }
 
 fn handle_payload_from_frame_acquire_thread(
@@ -139,11 +138,7 @@ fn handle_payload_from_frame_acquire_thread(
                 }),
             camera_file_transfer_in_progress: false,
         }) => {
-            let model = if radiometry_enabled {
-                "lepton3.5"
-            } else {
-                "lepton3"
-            };
+            let model = if radiometry_enabled { "lepton3.5" } else { "lepton3" };
             let header = format!(
                 "ResX: 160\n\
                         ResX: 160\n\
@@ -158,9 +153,7 @@ fn handle_payload_from_frame_acquire_thread(
             for (_, _, stream) in sockets.iter_mut().filter(|(_, use_wifi, stream)| {
                 stream.is_some() && !use_wifi && !stream.as_ref().unwrap().sent_header
             }) {
-                let stream = stream
-                    .as_mut()
-                    .expect("Never fails, because we filtered already.");
+                let stream = stream.as_mut().expect("Never fails, because we filtered already.");
                 if stream.write_all(header.as_bytes()).is_err() {
                     warn!("Failed sending header info");
                 }
@@ -190,11 +183,7 @@ fn handle_payload_from_frame_acquire_thread(
                     if !sent {
                         warn!(
                             "Send to {} failed",
-                            if *use_wifi {
-                                "tc2-frames server"
-                            } else {
-                                address
-                            }
+                            if *use_wifi { "tc2-frames server" } else { address }
                         );
                         let _ = stream.take().expect("Never fails").shutdown().is_ok();
                     }
@@ -202,21 +191,19 @@ fn handle_payload_from_frame_acquire_thread(
             }
             let e = s.elapsed().as_secs_f32();
             if e > 0.1 {
-                info!("socket send took {}s", e);
+                info!("socket send took {e}s");
             }
             if let Some(telemetry) = telemetry {
                 if let Some(prev_frame_num) = prev_frame_num {
-                    if !telemetry.ffc_in_progress {
-                        if telemetry.frame_num != *prev_frame_num + 1 {
-                            // NOTE: Frames can be missed when the raspberry pi
-                            //  blocks the thread with the
-                            //  unix socket in `thermal-recorder`.
-                            debug!(
-                                "Missed {} frames after {}s on",
-                                telemetry.frame_num - (*prev_frame_num + 1),
-                                telemetry.msec_on as f32 / 1000.0
-                            );
-                        }
+                    if !telemetry.ffc_in_progress && telemetry.frame_num != *prev_frame_num + 1 {
+                        // NOTE: Frames can be missed when the raspberry pi
+                        //  blocks the thread with the
+                        //  unix socket in `thermal-recorder`.
+                        debug!(
+                            "Missed {} frames after {}s on",
+                            telemetry.frame_num - (*prev_frame_num + 1),
+                            telemetry.msec_on as f32 / 1000.0
+                        );
                     }
                 }
                 *prev_frame_num = Some(telemetry.frame_num);
@@ -240,11 +227,7 @@ fn handle_payload_from_frame_acquire_thread(
                     {
                         info!(
                             "Shutting down socket '{}'",
-                            if *use_wifi {
-                                "tc2-frames server"
-                            } else {
-                                address
-                            }
+                            if *use_wifi { "tc2-frames server" } else { address }
                         );
                         let _ = stream.take().unwrap().shutdown().is_ok();
                     }
